Refactored user routes to user controllers, added user avatar and git ignore file.
