+++ server/.gitignore
@@ -1,0 +1,3 @@
+# Ignore folder name node modules
+node_modules/
+userAvatars/
