+++ server/api/helpers/helper.js
@@ -1,2 +1,1 @@
-exports.dateToString = date => new Date(date).toISOString();
 exports.privateKey = 'secretkey';
